Cleanup
1. Add day 7 input
2. Cleanup previous solvers
3. Create "Day00" empty starting solver class to clone each day from

+++ Solvers/Day03.cs
@@ -90,7 +90,7 @@ namespace Solvers
 
         private int GetPriorityForChar(char c)
         {
-            int charAsInt = (int)c;
+            int charAsInt = c;
             if (charAsInt >= 97)
             {
                 // lower case letters
